changes modification

+++ src/main/java/StatsService.java
@@ -1,5 +1,5 @@
 public class StatsService {
-    public int amountOfSales(long[] sales) {
+    public long amountOfSales(long[] sales) {
         //Сумма продаж за все месяцы
         long amount = 0;
         for (int i = 0; i < sales.length; i++) {
@@ -8,7 +8,7 @@ public class StatsService {
         return (int) amount;
     }
 
-    public int averageAmount(long[] sales) {
+    public long averageAmount(long[] sales) {
         //Среднее значение за все месяцы
         long average = 0;
         long amount = amountOfSales(sales);
@@ -18,7 +18,7 @@ public class StatsService {
         return (int) average;
     }
 
-    public int maxSales(long[] sales) {
+    public long maxSales(long[] sales) {
         int maxMonth = 0; // номер месяца с максимальными продажами среди просмотренных ранее
 
         for (int i = 0; i < sales.length; i++) {
@@ -30,7 +30,7 @@ public class StatsService {
         return maxMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
 
-    public int minSales(long[] sales) {
+    public long minSales(long[] sales) {
         int minMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
 
         for (int i = 0; i < sales.length; i++) {
@@ -42,7 +42,7 @@ public class StatsService {
         return minMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
 
-    public int salesBelowAverage(long[] sales) {
+    public long salesBelowAverage(long[] sales) {
         long average = averageAmount(sales);
         int belowAverageMoth = 0;
         for (int i = 0; i < sales.length; i++) {
@@ -53,7 +53,7 @@ public class StatsService {
         return belowAverageMoth;
     }
 
-    public int salesAboveverage(long[] sales) {
+    public long salesAboveverage(long[] sales) {
         long average = averageAmount(sales);
         int aboveAverageMoth = 0;
         for (int i = 0; i < sales.length; i++) {
